+++ app/models/activity.rb
@@ -153,4 +153,8 @@ class Activity < ApplicationRecord
   def transactions_total
     transactions.sum(:value)
   end
+
+  def actual_total_for_report_financial_quarter(report:)
+    transactions.where(report: report, date: report.created_at.all_quarter).sum(:value)
+  end
 end