+++ spec/services/export_activity_to_csv_spec.rb
@@ -2,12 +2,13 @@ require "rails_helper"
 require "csv"
 
 RSpec.describe ExportActivityToCsv do
-  let(:project) { create(:project_activity) }
+  let(:project) { create(:project_activity, :with_report) }
+  let(:report) { Report.find_by(fund: project.associated_fund, organisation: project.organisation) }
 
   describe "#call" do
     it "creates a CSV line representation of the Activity" do
       activity_presenter = ActivityPresenter.new(project)
-      result = ExportActivityToCsv.new(activity: project).call
+      result = ExportActivityToCsv.new(activity: project, report: report).call
 
       expect(result).to eq([
         activity_presenter.identifier,
@@ -24,7 +25,7 @@ RSpec.describe ExportActivityToCsv do
         activity_presenter.recipient_country,
         activity_presenter.aid_type,
         activity_presenter.level,
-        activity_presenter.transactions_total,
+        activity_presenter.actual_total_for_report_financial_quarter(report: report),
         activity_presenter.link_to_roda,
       ].to_csv)
     end
@@ -35,7 +36,7 @@ RSpec.describe ExportActivityToCsv do
       travel_to(Date.parse("1 April 2020")) do
         report = Report.new
 
-        headers = ExportActivityToCsv.new.headers(report: report)
+        headers = ExportActivityToCsv.new(activity: build(:activity), report: report).headers
 
         expect(headers).to include "Q1 2020-2021 actuals"
       end