+++ CHANGELOG.md
@@ -246,6 +246,9 @@
   Mapping `programme_status` to `IATI_status` included. Schema migration to replace
   `form_state` at `status` step for `programme_status` step
 - `programme status` form step is not shown for level A activities
+- Scope the total transaction value for an Activity to a Report and that Report's date
+  range, and call the result `actual_total_for_report_financial_quarter`. Add this
+  value to the Report CSV.
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-13...HEAD
 [release-13]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-12...release-13